core/requirements.txt
djangorestframework
djangorestframework-simplejwt
mysqlclient
django-widget-tweaks